Add Punditize support for namespaced models and policy fallbacks
