+++ spec/example_app/app/policies/admin/product_policy.rb
@@ -1,0 +1,37 @@
+module Admin
+  class ProductPolicy < ApplicationPolicy
+    class Scope < Scope
+      def resolve
+        scope.all
+      end
+    end
+
+    def index?
+      true
+    end
+
+    def show?
+      scope.where(id: record.id).exists?
+    end
+
+    def create?
+      true
+    end
+
+    def new?
+      create?
+    end
+
+    def update?
+      true
+    end
+
+    def edit?
+      update?
+    end
+
+    def destroy?
+      true
+    end
+  end
+end
